==> 4. CNN-LSTM.ipynb <==
# 4. CNN-LSTM

import numpy as np
import pandas as pd
import time
import os
os.environ["CUDA_VISIBLE_DEVICES"] = "0"

from tensorflow.python.client import device_lib
print(device_lib.list_local_devices())

import matplotlib.pyplot as plt
import tensorflow as tf
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Flatten, Conv1D, MaxPooling1D, ReLU, LSTM
from tensorflow.keras.layers import TimeDistributed
from focal_loss import BinaryFocalLoss

from sklearn.metrics import roc_curve

feature = ['flux',
           'zero_month',
           'zero_week',
           'continuous_zero',
           'diff_week_day',
           'diff_month',
           'std_week',
           'diff_std',
           'rec_day',
           'rec_week',
           'rec_2week',
           'rec_month',
           'rec_day2',
           'rec_week2',
           'rec_2week2',
           'rec_month2', ]

total_df = pd.read_csv("../final_data.csv")
total_data = total_df.dropna().loc[:, ['label'] + feature].to_numpy()

window_n = 5
stride = 12
window_size = 24

normal_data = []
abnormal_data = []
where_abnormal = np.argwhere(total_data[:,0]==1)

# normal_data 추출 과정
for i in range(total_data.shape[0] - (window_size + (window_n - 1) * stride)):
  # 무작위성 부여
  if np.random.rand() > 0.30:
    continue
  
  sliced = []
  for w_n in range(window_n):
    # sliced 에 window_n번 window_size 행만큼 자른 데이터 append
    # sliced 에 5번 24x17 추가
    sliced.append(total_data[(w_n * stride) + i : (w_n * stride) + i + window_size])
  
  sliced = np.array(sliced)

  # 자른 데이터 label이 모두 정상이면 normal 추가
  if np.all(sliced[:,:,0] == 0):
    normal_data.append(sliced[:, :, 1:])

normal_data = np.array(normal_data).astype(np.float32)



# abnormal_data 추출 과정
for ab in where_abnormal:
    # 이전 24 + 12*4 시간의 data를 보고 진단 == 72시간
    # 그냥 ab 로 하면 list --> ab[0]
    i = ab[0] + 1 - window_size - stride * (window_n - 1)
    if (i < 0):
        continue
    
    sliced = []
    for w_n in range(window_n):
        sliced.append(total_data[i + (w_n * stride) : i + (w_n * stride) + window_size, 1:])
    
    sliced = np.array(sliced)
    abnormal_data.append(sliced)

#abnormal_data : 고장 직전 72시간의 데이터를 포함
abnormal_data = np.array(abnormal_data).astype(np.float32)

# Index Shuffle
random_idx = np.arange(normal_data.shape[0])
np.random.shuffle(random_idx)
normal_data = normal_data[random_idx]

random_idx = np.arange(abnormal_data.shape[0])
np.random.shuffle(random_idx)
abnormal_data = abnormal_data[random_idx]


cut = normal_data.shape[0]//10

# Training Dataset
training_x = normal_data[:cut]
training_y = [[1, 0] for _ in range(cut)]

# abnormal data oversampling 200 times
duplicate_weight = 200
for _ in range(duplicate_weight):
  training_x = np.concatenate([training_x, abnormal_data[:abnormal_data.shape[0]//2]], axis = 0)


training_y = training_y + [[0, 1] for _ in range(training_x.shape[0]-len(training_y))]
training_y = np.array(training_y)


# Test Dataset
test_x = normal_data[cut:cut*2]
test_y = [[1, 0] for _ in range(cut)]

test_x = np.concatenate([test_x, abnormal_data[abnormal_data.shape[0]//2:]], axis = 0)
test_y = test_y + [[0, 1] for _ in range(test_x.shape[0]-len(test_y))]

test_y = np.array(test_y)


# Dataset Labeling
# normal : [1,0]
# abnormal : [0,1]

print(training_x.shape)
print(training_y.shape)
print(test_x.shape)
print(test_y.shape)

print("training data 중 abnormal : ", sum(training_y[:,1]))

- Cross Entropy Loss

l2param=0.001

model = Sequential()
model.add(TimeDistributed(
  Conv1D(32, 4, padding = "valid",
         kernel_regularizer=tf.keras.regularizers.l2(l2param))
  ))
model.add(TimeDistributed(ReLU()))
model.add(TimeDistributed(MaxPooling1D(2)))
model.add(TimeDistributed(Flatten()))

model.add(LSTM(units=256,
              kernel_regularizer=tf.keras.regularizers.l2(l2param)
              ))
model.add(Dense(2, activation='sigmoid',
                kernel_regularizer=tf.keras.regularizers.l2(l2param)))

model.compile(loss='binary_crossentropy', optimizer=Adam(learning_rate=0.0005), metrics=['accuracy'])

start = time.time()
history = model.fit(training_x, training_y, epochs=10, batch_size=256, validation_split=0.1)
end = time.time()

pred = model.predict(test_x, batch_size=256)[:,1]
test_ground_truth = test_y[:,1]
fpr, tpr, thresholds = roc_curve(test_ground_truth, pred)

# test_y
# normal : [1,0]
# abnormal : [0,1]

for i in range(len(fpr)):
        if fpr[i] > 0.001: #0.1%
            # fpr 이 thresholds 값 이상이 될 때 그 index 를 찾음
            if (i > 0): i -= 1
            break

predicted = pred >= thresholds[i]
TP = int(sum(np.logical_and(test_ground_truth==1,  predicted==1)))
TN = int(sum(np.logical_and(test_ground_truth==0,  predicted==0)))
FP = int(sum(np.logical_and(test_ground_truth==0,  predicted==1)))
FN = int(sum(np.logical_and(test_ground_truth==1,  predicted==0)))

Precision = TP/(TP+FP)
Recall = TP/(TP+FN)

print(TP, TN, FP, FN)

print("precision: ", Precision)
print("recall: ", Recall)
print("F-measure:", (2*Precision*Recall)/(Precision+Recall+1e-7))
print("MCC:", (TP*TN-FP*FN)/(np.sqrt(float((TP+FN)*(TP+FP)*(TN+FP)*(TN+FN)))+1e-7))
print("train time : ", end-start)

- Focal Loss

l2param=0.001

gamm = [0.5, 1, 1.5, 2.0]
for gam in gamm :
    
    model = Sequential()
    model.add(TimeDistributed(
      Conv1D(32, 4, padding = "valid",
            kernel_regularizer=tf.keras.regularizers.l2(l2param))
      ))
    model.add(TimeDistributed(ReLU()))
    model.add(TimeDistributed(MaxPooling1D(2)))
    model.add(TimeDistributed(Flatten()))

    model.add(LSTM(units=256,
                  kernel_regularizer=tf.keras.regularizers.l2(l2param)
                  ))
    model.add(Dense(2, activation='sigmoid',
                    kernel_regularizer=tf.keras.regularizers.l2(l2param)))
    model.compile(loss=BinaryFocalLoss(gamma=gam), optimizer=Adam(learning_rate=0.0005), metrics=['accuracy'])
    model.fit(training_x, training_y, epochs=10, batch_size=256, verbose=0, validation_split=0.1)

    print('gamma = {}'.format(gam))

    pred = model.predict(test_x, batch_size=256)[:,1]
    test_ground_truth = test_y[:,1]
    fpr, tpr, thresholds = roc_curve(test_ground_truth, pred)

    # test_y
    # normal : [1,0]
    # abnormal : [0,1]

    for i in range(len(fpr)):
            if fpr[i] > 0.001: #0.1%
                # fpr 이 thresholds 값 이상이 될 때 그 index 를 찾음
                if (i > 0): i -= 1
                break

    predicted = pred >= thresholds[i]
    TP = int(sum(np.logical_and(test_ground_truth==1,  predicted==1)))
    TN = int(sum(np.logical_and(test_ground_truth==0,  predicted==0)))
    FP = int(sum(np.logical_and(test_ground_truth==0,  predicted==1)))
    FN = int(sum(np.logical_and(test_ground_truth==1,  predicted==0)))

    Precision = TP/(TP+FP)
    Recall = TP/(TP+FN)

    print(TP, TN, FP, FN)

    print("precision: ", Precision)
    print("recall: ", Recall)
    print("F-measure:", (2*Precision*Recall)/(Precision+Recall+1e-7))
    print("MCC:", (TP*TN-FP*FN)/(np.sqrt(float((TP+FN)*(TP+FP)*(TN+FP)*(TN+FN)))+1e-7))
    print('='*20)
